add login feat

+++ src/utils/logout.ts
@@ -1,4 +1,7 @@
-export default function logout() {
+export default function logout(code?: number) {
+  if (code && code !== 401) {
+    return;
+  }
   localStorage.clear();
   window.location.reload();
 }
